+++ Tests/AVFoundationCollectorTests/mock/AccessLogProviderMock.swift
@@ -1,0 +1,11 @@
+import Foundation
+@testable import AVFoundationCollector
+@testable import CoreCollector
+
+class AccessLogProviderMock: AccessLogProvider {
+    var events: [AccessLogDto]?
+
+    func getEvents() -> [AccessLogDto]? {
+        events
+    }
+}